fix: rename LoggerProvider component
